don't overwrite config file on install

+++ dump1090-fa-autogain.sh
@@ -5,6 +5,9 @@
 mkdir -p /usr/local/bin
 cat >/usr/local/bin/dump1090-fa-autogain <<"EOF"
 #!/bin/bash
+low=1.0
+high=5.0
+ga=(0.0 0.9 1.4 2.7 3.7 7.7 8.7 12.5 14.4 15.7 16.6 19.7 20.7 22.9 25.4 28.0 29.7 32.8 33.8 36.4 37.2 38.6 40.2 42.1 43.4 43.9 44.5 48.0 49.6 -10)
 source /etc/default/dump1090-fa-autogain
 strong=$(awk "$(cat /run/dump1090-fa/stats.json | grep total | sed 's/.*accepted":\[\([0-9]*\).*strong_signals":\([0-9]*\).*/BEGIN {printf "%.3f" , \2 * 100 \/ \1}/')")
 
@@ -46,7 +49,9 @@ echo "$action gain to $gain (${strong}% messages >-3dB)"
 EOF
 chmod a+x /usr/local/bin/dump1090-fa-autogain
 
-cat >/etc/default/dump1090-fa-autogain <<"EOF"
+config_file=/etc/default/dump1090-fa-autogain
+if ! [ -f $config_file ]; then
+	cat >/etc/default/dump1090-fa-autogain <<"EOF"
 #!/bin/bash
 
 low=1.0
@@ -54,6 +59,7 @@ high=5.0
 
 ga=(0.0 0.9 1.4 2.7 3.7 7.7 8.7 12.5 14.4 15.7 16.6 19.7 20.7 22.9 25.4 28.0 29.7 32.8 33.8 36.4 37.2 38.6 40.2 42.1 43.4 43.9 44.5 48.0 49.6 -10)
 EOF
+fi
 
 cat >/etc/cron.d/dump1090-fa-autogain <<"EOF"
 45 2 * * * root /bin/bash /usr/local/bin/dump1090-fa-autogain
